Add The Odin Project Ruby and Rails courses to the curriculum
The Odin Project is an online platform for learning web development. It features seven courses, from what a web developer does, all the way to how to get hired. In between those courses you have a Ruby course and a Rails course. 
The idea of the platform is to learn by doing, you'll usually have assignments to read and then a hands-on project to solidify what you just learned. You build a lot of stuff, in the ruby course for example, you build a Mastermind game, Tic Tac Toe game, Chess, among others.